Split the mega `CollectionDiffing.swift` up.
