+++ Sources/CollectionDiffing.swift
@@ -1,96 +1,6 @@
 import Foundation
 import Result
 
-/// Represents a snapshot of a collection.
-///
-/// The changeset associated with the first snapshot received by any given observation
-/// is generally irrelevant. Observations should special case the first snapshot as a
-/// complete replacement, or whatever semantic that fits their purpose.
-///
-/// `Snapshot` does not include a reference to the previous snapshot. If the
-/// history is relevant for delta consumption, it is up to the observers to cache the
-/// history on their own.
-///
-/// While the protocol does not bind the `Changeset` associated type with any constraint,
-/// it is expected to be either a `Changeset`, or a composite of `Changeset`s for a nested
-/// collection.
-public struct Snapshot<Elements: Collection, Changeset>: SnapshotProtocol {
-	public let elements: Elements
-	public let changeset: Changeset
-
-	public init(elements: Elements, changeset: Changeset) {
-		(self.elements, self.changeset) = (elements, changeset)
-	}
-}
-
-public protocol SnapshotProtocol {
-	associatedtype Elements: Collection
-	associatedtype Changeset
-
-	var elements: Elements { get }
-	var changeset: Changeset { get }
-}
-
-/// Represents an atomic batch of changes made to a collection.
-///
-/// A `Changeset` represents changes as **offsets** of elements. You may
-/// subscript a collection of zero-based indexing with the offsets, e.g. `Array`. You must
-/// otherwise convert the offsets into indices before subscripting.
-public struct Changeset {
-	/// Represents the context of a move operation applied to a collection.
-	public struct Move {
-		public let source: Int
-		public let isMutated: Bool
-
-		public init(source: Int, isMutated: Bool) {
-			(self.source, self.isMutated) = (source, isMutated)
-		}
-	}
-
-	/// The offsets of inserted elements **after** the removals were applied.
-	///
-	/// - important: To obtain the actual index, you must apply
-	///              `Collection.index(self:_:offsetBy:)` on the current snapshot, the
-	///              start index and the offset.
-	public var inserts = IndexSet()
-
-	/// The offsets of removed elements **prior to** any changes being applied.
-	///
-	/// - important: To obtain the actual index, you must apply
-	///              `Collection.index(self:_:offsetBy:)` on the previous snapshot, the
-	///              start index and the offset.
-	public var removals = IndexSet()
-
-	/// The offsets of position-invariant mutations.
-	///
-	/// `mutations` only implies an invariant relative position. The actual indexes can
-	/// be different, depending on the collection type.
-	///
-	/// If an element has both changed and moved, it would be included in `moves` with an
-	/// asserted mutation flag.
-	///
-	/// - important: To obtain the actual index, you must apply
-	///              `Collection.index(self:_:offsetBy:)` on the relevant snapshot, the
-	///              start index and the offset.
-	public var mutations = IndexSet()
-
-	/// The offset pairs of moves.
-	///
-	/// The offset pairs are recorded as a `Dictionary` keyed by the destination offset,
-	/// with the source offset and a mutation flag as the associated value.
-	///
-	/// - important: To obtain the actual index, you must apply
-	///              `Collection.index(self:_:offsetBy:)` on the relevant snapshot, the
-	///              start index and the offset.
-	public var moves = [Int: Move]()
-
-	public init() {}
-
-	public init<C: Collection>(initial: C) {
-		inserts = IndexSet(integersIn: 0 ..< Int(initial.count))
-	}
-}
-
 /// The comparison strategies used by the collection diffing operators on collections
 /// that contain `Hashable` objects.
 public struct ObjectDiffStrategy {
@@ -211,7 +121,7 @@ extension Signal where Value: Collection, Value.Indices.Iterator.Element == Valu
 					let changeset: Changeset
 
 					if let previous = previous {
-						changeset = Value.diff(previous: previous, current: elements, identifier: identifier, areEqual: areEqual)
+						changeset = Changeset(previous: previous, current: elements, identifier: identifier, areEqual: areEqual)
 					} else {
 						changeset = Changeset(initial: elements)
 					}
@@ -395,166 +305,3 @@ extension PropertyProtocol where Value: Collection, Value.Indices.Iterator.Eleme
 		return lift { $0.diff(identifier: identifier, areEqual: areEqual) }
 	}
 }
-
-// MARK: - Implementation details
-
-// The key equality implies only referential equality. But the value equality of the
-// uniquely identified element across snapshots is uncertain. It is pretty common to diff
-// elements with constant unique identifiers but changing contents. For example, we may
-// have an array of `Conversation`s, identified by the backend ID, that is constantly
-// updated with the latest messages pushed from the backend. So our diffing algorithm
-// must have an additional mean to test elements for value equality.
-
-private final class DiffEntry {
-	var occurenceInOld = 0
-	var occurenceInNew = 0
-	var locationInOld: Int?
-}
-
-private enum DiffReference {
-	case remote(Int)
-	case table(DiffEntry)
-}
-
-// FIXME: Swift 4 Associated type constraints
-// extension Collection {
-extension Collection where Index == Indices.Iterator.Element {
-	// @testable
-	internal static func diff<Identifier: Hashable>(
-		previous: Self,
-		current: Self,
-		identifier: (Iterator.Element) -> Identifier,
-		areEqual: (Iterator.Element, Iterator.Element) -> Bool
-	) -> Changeset {
-		switch Self.self {
-		case is Array<Iterator.Element>.Type:
-			fallthrough
-		case is ContiguousArray<Iterator.Element>.Type:
-			// The standard library does not copy the content when `ContiguousArray.init`
-			// is passed an `Array` backed by a native buffer or a `ContiguousArray`. So
-			// we exploit this optimization to obtain an unsafe buffer pointer.
-			//
-			// `ArraySlice` is not supported at the moment, since it may have a non-zero
-			// start index that needs to be taken care of.
-			return ContiguousArray(previous).withUnsafeBufferPointer { previousBuffer in
-				return ContiguousArray(current).withUnsafeBufferPointer { currentBuffer in
-					return _diff(previous: previousBuffer,
-					             current: currentBuffer,
-					             identifier: identifier,
-					             areEqual: areEqual)
-				}
-			}
-
-		default:
-			return _diff(previous: previous,
-			             current: current,
-			             identifier: identifier,
-			             areEqual: areEqual)
-		}
-	}
-
-	private static func _diff<View: Collection, Identifier: Hashable>(
-		previous: View,
-		current: View,
-		identifier: (Iterator.Element) -> Identifier,
-		areEqual: (Iterator.Element, Iterator.Element) -> Bool
-	) -> Changeset where View.Iterator.Element == Iterator.Element, View.Index == View.Indices.Iterator.Element {
-		var table: [Identifier: DiffEntry] = Dictionary(minimumCapacity: Int(current.count))
-		var oldReferences: [DiffReference] = []
-		var newReferences: [DiffReference] = []
-
-		oldReferences.reserveCapacity(Int(previous.count))
-		newReferences.reserveCapacity(Int(current.count))
-
-		func tableEntry(for identifier: Identifier) -> DiffEntry {
-			if let index = table.index(forKey: identifier) {
-				return table[index].value
-			}
-
-			let entry = DiffEntry()
-			table[identifier] = entry
-			return entry
-		}
-
-		// Pass 1: Scan the new snapshot.
-		for element in current {
-			let key = identifier(element)
-			let entry = tableEntry(for: key)
-
-			entry.occurenceInNew += 1
-			newReferences.append(.table(entry))
-		}
-
-		// Pass 2: Scan the old snapshot.
-		for (offset, index) in previous.indices.enumerated() {
-			let key = identifier(previous[index])
-			let entry = tableEntry(for: key)
-
-			entry.occurenceInOld += 1
-			entry.locationInOld = offset
-			oldReferences.append(.table(entry))
-		}
-
-		// Pass 3: Single-occurence lines
-		for newPosition in newReferences.startIndex ..< newReferences.endIndex {
-			switch newReferences[newPosition] {
-			case let .table(entry):
-				if entry.occurenceInNew == 1 && entry.occurenceInNew == entry.occurenceInOld {
-					let oldPosition = entry.locationInOld!
-					newReferences[newPosition] = .remote(oldPosition)
-					oldReferences[oldPosition] = .remote(newPosition)
-				}
-
-			case .remote:
-				break
-			}
-		}
-
-		var changeset = Changeset()
-
-		// Final Pass: Compute diff.
-		for position in oldReferences.indices {
-			switch oldReferences[position] {
-			case .table:
-				// Deleted
-				changeset.removals.insert(position)
-
-			case let .remote(newPosition):
-				let previousIndex = previous.index(previous.startIndex, offsetBy: View.IndexDistance(position))
-				let currentIndex = current.index(current.startIndex, offsetBy: View.IndexDistance(newPosition))
-				let areEqual = areEqual(previous[previousIndex], current[currentIndex])
-
-				// If the move is only caused by a deletion earlier on, it is still
-				// considered in place.
-				let isInPlace = newPosition == position || position - newPosition == changeset.removals.count(in: 0 ..< position)
-
-				switch (areEqual, isInPlace) {
-				case (false, true):
-					changeset.mutations.insert(position)
-
-				case (_, false):
-					changeset.moves[newPosition] = Changeset.Move(source: position, isMutated: !areEqual)
-
-				case (true, true):
-					break
-				}
-			}
-		}
-
-		for position in newReferences.indices {
-			if case .table = newReferences[position] {
-				changeset.inserts.insert(position)
-			}
-		}
-
-		return changeset
-	}
-}
-
-#if !swift(>=3.2)
-	extension SignedInteger {
-		fileprivate init<I: SignedInteger>(_ integer: I) {
-			self.init(integer.toIntMax())
-		}
-	}
-#endif
